requirements.txt
allure-pytest
pytest
requests